completed unit testing

+++ UnitTesting/app/src/main/java/com/ryvk/unittesting/util/Calculator.kt
@@ -4,8 +4,8 @@ class Calculator {
     fun add(a: Int, b: Int): Int = a + b
     fun subtract(a: Int, b: Int): Int = a - b
     fun multiply(a: Int, b: Int): Int = a * b
-    fun divide(a: Int, b: Int): Int {
-        require(b != 0) { "Cannot divide by zero" }
+    fun divide(a: Double, b: Double): Double {
+        require(b != 0.00) { "Cannot divide by zero" }
         return a / b
     }
